Migrating HatPtr to use schema.
