+++ hat/hat/src/main/java/hat/ifacemapper/Schema.java
@@ -20,13 +20,15 @@ public class Schema<T extends Buffer> {
     final public IfaceType rootIfaceType;
     public Class<T> iface;
 
-    public static abstract  class SchemaNode  {
+    public static abstract class SchemaNode {
         public static final class Padding extends FieldNode {
             int len;
+
             Padding(IfaceType parent, int len) {
-                super(parent, AccessorInfo.Key.NONE, "pad"+len);
+                super(parent, AccessorInfo.Key.NONE, "pad" + len);
                 this.len = len;
             }
+
             @Override
             public void toText(String indent, Consumer<String> stringConsumer) {
                 stringConsumer.accept(indent + "padding " + len + " bytes");
@@ -40,15 +42,15 @@ public class Schema<T extends Buffer> {
     }
 
 
-    public T allocate(MethodHandles.Lookup lookup,BufferAllocator bufferAllocator, int... boundLengths) {
+    public T allocate(MethodHandles.Lookup lookup, BufferAllocator bufferAllocator, int... boundLengths) {
         BoundSchema<?> boundSchema = new BoundSchema<>(this, boundLengths);
-        return (T) boundSchema.allocate(lookup,bufferAllocator);
+        return (T) boundSchema.allocate(lookup, bufferAllocator);
     }
 
-    public static <T extends Buffer> Schema<T> of(Class<T> iface,  Consumer<IfaceType> parentFieldConsumer) {
-        var struct = new IfaceType.Struct(null, (Class<MappableIface>)(Object)iface); // why the need for this?
+    public static <T extends Buffer> Schema<T> of(Class<T> iface, Consumer<IfaceType> parentFieldConsumer) {
+        var struct = new IfaceType.Struct(null, (Class<MappableIface>) (Object) iface); // why the need for this?
         parentFieldConsumer.accept(struct);
-        return new Schema<>(iface,struct);
+        return new Schema<>(iface, struct);
     }
 
     public void toText(Consumer<String> stringConsumer) {
@@ -77,49 +79,48 @@ public class Schema<T extends Buffer> {
             this.iface = iface;
         }
 
-        IfaceType getChild(Class<?> iface){
+        IfaceType getChild(Class<?> iface) {
             Optional<IfaceType> ifaceTypeNodeOptional = ifaceTypes
                     .stream()
-                    .filter(n->n.iface.equals(iface))
+                    .filter(n -> n.iface.equals(iface))
                     .findFirst();
-            if (ifaceTypeNodeOptional.isPresent()){
+            if (ifaceTypeNodeOptional.isPresent()) {
                 return ifaceTypeNodeOptional.get();
-            }else {
+            } else {
                 throw new IllegalStateException("no supported iface type");
             }
         }
 
         public void visitTypes(int depth, Consumer<IfaceType> ifaceTypeNodeConsumer) {
-            ifaceTypes.forEach(t->t.visitTypes(depth+1,ifaceTypeNodeConsumer));
+            ifaceTypes.forEach(t -> t.visitTypes(depth + 1, ifaceTypeNodeConsumer));
             ifaceTypeNodeConsumer.accept(this);
         }
 
-        public GroupLayout getBoundGroupLayout( BoundSchema.BoundSchemaNode parentBoundSchemaNode){
+        public GroupLayout getBoundGroupLayout(BoundSchema.BoundSchemaNode parentBoundSchemaNode) {
 
-            BoundSchema.BoundSchemaNode child =parentBoundSchemaNode.createChild(this);
+            BoundSchema.BoundSchemaNode<?> child = parentBoundSchemaNode.createChild(this);
             this.fields.forEach(fieldNode -> {
                         if (fieldNode instanceof SchemaNode.Padding field) {
                             child.bind(field, MemoryLayout.paddingLayout(field.len));
-                        }else if (fieldNode instanceof FieldNode.AddressField field) {
+                        } else if (fieldNode instanceof FieldNode.AddressField field) {
                             child.bind(field, field.parent.getBoundLayout(field.type, child).withName(field.name));
-                        }else if (fieldNode instanceof FieldNode.ArrayLen field) {
+                        } else if (fieldNode instanceof FieldNode.ArrayLen field) {
                             child.bind(field, field.parent.getBoundLayout(field.type, child).withName(field.name));
-                        }else if (fieldNode instanceof FieldNode.AtomicField field) {
+                        } else if (fieldNode instanceof FieldNode.AtomicField field) {
                             child.bind(field, field.parent.getBoundLayout(field.type, child).withName(field.name));
-                        }else if (fieldNode instanceof FieldNode.IfaceField field) {
-                            child.bind(field, field.parent.getBoundLayout(field.type.iface, child).withName(field.name));
-                        }else if (fieldNode instanceof FieldNode.PrimitiveField field) {
+                        } else if (fieldNode instanceof FieldNode.IfaceField field) {
+                            child.bind(field, field.parent.getBoundLayout(field.ifaceType.iface, child).withName(field.name));
+                        } else if (fieldNode instanceof FieldNode.PrimitiveField field) {
                             child.bind(field, field.parent.getBoundLayout(field.type, child).withName(field.name));
-                        }else if (fieldNode instanceof FieldNode.IfaceFixedArray field) {
+                        } else if (fieldNode instanceof FieldNode.IfaceFixedArray field) {
                             child.bind(field, MemoryLayout.sequenceLayout(field.len,
-                                    field.parent.getBoundLayout(field.type.iface, child).withName(field.type.iface.getSimpleName())
+                                    field.parent.getBoundLayout(field.ifaceType.iface, child).withName(field.ifaceType.iface.getSimpleName())
                             ).withName(field.name));
-                        }else if (fieldNode instanceof FieldNode.PrimitiveFixedArray field) {
+                        } else if (fieldNode instanceof FieldNode.PrimitiveFixedArray field) {
                             child.bind(field, MemoryLayout.sequenceLayout(field.len,
                                     field.parent.getBoundLayout(field.type, child).withName(field.type.getSimpleName())
                             ).withName(field.name));
-                        }else if (fieldNode instanceof FieldNode.IfaceFieldControlledArray field) {
-
+                        } else if (fieldNode instanceof FieldNode.IfaceFieldControlledArray field) {
                             // To determine the actual 'array' size we multiply the contributing dims by the stride .
                             int size = field.stride; //usually 1 but developer can define.
                             for (int i = 0; i < field.contributingDims; i++) {
@@ -127,21 +128,19 @@ public class Schema<T extends Buffer> {
                             }
 
                             child.bind(field, MemoryLayout.sequenceLayout(size,
-                                    field.parent.getBoundLayout(field.type.iface, child).withName(field.type.iface.getSimpleName())
+                                    field.parent.getBoundLayout(field.ifaceType.iface, child).withName(field.ifaceType.iface.getSimpleName())
                             ).withName(field.name));
-                        }else if (fieldNode instanceof FieldNode.PrimitiveFieldControlledArray field){
+                        } else if (fieldNode instanceof FieldNode.PrimitiveFieldControlledArray field) {
+                            // To determine the actual 'array' size we multiply the contributing dims by the stride .
+                            int size = field.stride; //usually 1 but developer can define.
+                            for (int i = 0; i < field.contributingDims; i++) {
+                                size *= child.takeArrayLen(); // this takes an arraylen and bumps the ptr
+                            }
 
-                                // To determine the actual 'array' size we multiply the contributing dims by the stride .
-                                int size = field.stride; //usually 1 but developer can define.
-                                for (int i = 0; i < field.contributingDims; i++) {
-                                    size *= child.takeArrayLen(); // this takes an arraylen and bumps the ptr
-                                }
-
-                                child.bind(field, MemoryLayout.sequenceLayout(size,
-                                        field.parent.getBoundLayout(field.type, child).withName(field.type.getSimpleName())
-                                ).withName(field.name));
-
-                        }else {
+                            child.bind(field, MemoryLayout.sequenceLayout(size,
+                                    field.parent.getBoundLayout(field.type, child).withName(field.type.getSimpleName())
+                            ).withName(field.name));
+                        } else {
                             throw new IllegalStateException("what is this?");
                         }
                     }
@@ -164,15 +163,15 @@ public class Schema<T extends Buffer> {
         MemoryLayout getBoundLayout(Class<?> type, BoundSchema.BoundSchemaNode boundSchemaNode) {
             if (type.isPrimitive()) {
                 return MapperUtil.primitiveToLayout(type);
-            }else if (MapperUtil.isMemorySegment(type)) {
+            } else if (MapperUtil.isMemorySegment(type)) {
                 return ValueLayout.ADDRESS;
             } else {
-               return  getChild(type).getBoundGroupLayout( boundSchemaNode);
+                return getChild(type).getBoundGroupLayout(boundSchemaNode);
             }
         }
 
         public IfaceType struct(String name, Consumer<IfaceType> parentSchemaNodeConsumer) {
-            parentSchemaNodeConsumer.accept(addIfaceTypeNode(new Struct(this, (Class<MappableIface>)MapperUtil.typeOf(iface, name))));
+            parentSchemaNodeConsumer.accept(addIfaceTypeNode(new Struct(this, (Class<MappableIface>) MapperUtil.typeOf(iface, name))));
             return this;
         }
 
@@ -185,10 +184,10 @@ public class Schema<T extends Buffer> {
             var key = AccessorInfo.Key.of(iface, name);
             var typeOf = MapperUtil.typeOf(iface, name);
             addField(MapperUtil.isMemorySegment(typeOf)
-                    ? new FieldNode.AddressField(this, key, (Class<MemorySegment>)typeOf, name)
+                    ? new FieldNode.AddressField(this, key, (Class<MemorySegment>) typeOf, name)
                     : MapperUtil.isMappableIface(typeOf)
-                       ? new FieldNode.IfaceField(this, key, this.getChild(typeOf), name)
-                       : new FieldNode.PrimitiveField(this, key, typeOf, name));
+                    ? new FieldNode.IfaceField(this, key, this.getChild(typeOf), name)
+                    : new FieldNode.PrimitiveField(this, key, typeOf, name));
             return this;
         }
 
@@ -204,10 +203,10 @@ public class Schema<T extends Buffer> {
 
         public IfaceType field(String name, Consumer<IfaceType> parentSchemaNodeConsumer) {
             AccessorInfo.Key fieldKey = AccessorInfo.Key.of(iface, name);
-            Class<MappableIface> fieldType = (Class<MappableIface>)MapperUtil.typeOf(iface, name);
-            IfaceType structOrUnion= MapperUtil.isStruct(fieldType) ? new Struct(this, fieldType) : new Union(this, fieldType);
+            Class<MappableIface> fieldType = (Class<MappableIface>) MapperUtil.typeOf(iface, name);
+            IfaceType structOrUnion = MapperUtil.isStruct(fieldType) ? new Struct(this, fieldType) : new Union(this, fieldType);
             addIfaceTypeNode(structOrUnion);
-            addField(new FieldNode.IfaceField(this, fieldKey, structOrUnion,name));
+            addField(new FieldNode.IfaceField(this, fieldKey, structOrUnion, name));
             parentSchemaNodeConsumer.accept(structOrUnion);
             return this;
         }
@@ -227,8 +226,8 @@ public class Schema<T extends Buffer> {
                     ? new Struct(this, structOrUnionType)
                     : new Union(this, structOrUnionType);
             addIfaceTypeNode(ifaceType);
-            addField(new FieldNode.IfaceField(this, fieldKey1, ifaceType,name1));
-            addField(new FieldNode.IfaceField(this, fieldKey2, ifaceType,name2));
+            addField(new FieldNode.IfaceField(this, fieldKey1, ifaceType, name1));
+            addField(new FieldNode.IfaceField(this, fieldKey2, ifaceType, name2));
 
             parentSchemaNodeConsumer.accept(ifaceType);
             return this;
@@ -243,16 +242,16 @@ public class Schema<T extends Buffer> {
 
         public IfaceType array(String name, int len) {
             AccessorInfo.Key arrayKey = AccessorInfo.Key.of(iface, name);
-            var typeof = MapperUtil.typeOf(iface,name);
+            var typeof = MapperUtil.typeOf(iface, name);
             addField(arrayKey.valueType().equals(ValueType.INTERFACE)
-                    ?new FieldNode.IfaceFixedArray( this,arrayKey,this.getChild(typeof),name, len)
-                    :new FieldNode.PrimitiveFixedArray(this, arrayKey, typeof, name, len));
+                    ? new FieldNode.IfaceFixedArray(this, arrayKey, this.getChild(typeof), name, len)
+                    : new FieldNode.PrimitiveFixedArray(this, arrayKey, typeof, name, len));
             return this;
         }
 
         public IfaceType array(String name, int len, Consumer<IfaceType> parentFieldConsumer) {
             AccessorInfo.Key arrayKey = AccessorInfo.Key.of(iface, name);
-            Class<MappableIface> structOrUnionType = (Class<MappableIface>)MapperUtil.typeOf(iface, name);
+            Class<MappableIface> structOrUnionType = (Class<MappableIface>) MapperUtil.typeOf(iface, name);
             IfaceType ifaceType = MapperUtil.isStruct(iface)
                     ? new Struct(this, structOrUnionType)
                     : new Union(this, structOrUnionType);
@@ -266,15 +265,15 @@ public class Schema<T extends Buffer> {
             AccessorInfo.Key arrayKey = AccessorInfo.Key.of(iface, name);
             var typeOf = MapperUtil.typeOf(iface, name);
             addField(arrayKey.valueType().equals(ValueType.INTERFACE)
-                    ?new FieldNode.IfaceFieldControlledArray(this, arrayKey, this.getChild(typeOf),name,  arrayLenFields, stride)
-                    :new FieldNode.PrimitiveFieldControlledArray(this, arrayKey, typeOf, name,  arrayLenFields, stride));
+                    ? new FieldNode.IfaceFieldControlledArray(this, arrayKey, this.getChild(typeOf), name, arrayLenFields, stride)
+                    : new FieldNode.PrimitiveFieldControlledArray(this, arrayKey, typeOf, name, arrayLenFields, stride));
             return this;
         }
 
         public static class ArrayBuildState {
             IfaceType ifaceType;
             List<FieldNode.ArrayLen> arrayLenFields;
-            int padding =0;
+            int padding = 0;
             int stride = 1;
 
             public IfaceType array(String name) {
@@ -285,12 +284,14 @@ public class Schema<T extends Buffer> {
                 this.stride = stride;
                 return this;
             }
+
             public ArrayBuildState pad(int padding) {
                 this.padding = padding;
                 var paddingField = new SchemaNode.Padding(ifaceType, padding);
                 ifaceType.addField(paddingField);
                 return this;
             }
+
             public IfaceType array(String name, Consumer<IfaceType> parentFieldConsumer) {
                 Class<MappableIface> arrayType = (Class<MappableIface>) MapperUtil.typeOf(this.ifaceType.iface, name);
                 IfaceType ifaceType = MapperUtil.isStruct(arrayType)
@@ -369,28 +370,13 @@ public class Schema<T extends Buffer> {
             this.key = key;
             this.name = name;
         }
+
         public abstract void toText(String indent, Consumer<String> stringConsumer);
 
-        public static abstract sealed class AbstractPrimitiveField extends FieldNode
-                permits  ArrayLen, AtomicField, PrimitiveArray, PrimitiveField {
-            public Class<?> type;
-            AbstractPrimitiveField(IfaceType parent, AccessorInfo.Key key, Class<?> type, String name) {
-                super(parent,key,  name);
-                this.type = type;
-            }
-        }
-
-        public static abstract sealed class AbstractIfaceField extends FieldNode
-                permits FieldNode.IfaceArray, FieldNode.IfaceField {
-            public IfaceType type;
-            AbstractIfaceField(IfaceType parent, AccessorInfo.Key key, IfaceType type, String name) {
-                super(parent,key, name);
-                this.type = type;
-            }
-        }
 
         public static final class AddressField extends FieldNode {
             Class<MemorySegment> type;
+
             AddressField(IfaceType parent, AccessorInfo.Key key, Class<MemorySegment> type, String name) {
                 super(parent, key, name);
                 this.type = type;
@@ -399,6 +385,16 @@ public class Schema<T extends Buffer> {
             @Override
             public void toText(String indent, Consumer<String> stringConsumer) {
                 stringConsumer.accept(indent + "address " + key + ":" + type);
+            }
+        }
+
+        public static abstract sealed class AbstractPrimitiveField extends FieldNode
+                permits ArrayLen, AtomicField, PrimitiveArray, PrimitiveField {
+            public Class<?> type;
+
+            AbstractPrimitiveField(IfaceType parent, AccessorInfo.Key key, Class<?> type, String name) {
+                super(parent, key, name);
+                this.type = type;
             }
         }
 
@@ -424,21 +420,8 @@ public class Schema<T extends Buffer> {
             }
         }
 
-        public static final class IfaceField extends AbstractIfaceField {
-
-            IfaceField(IfaceType parent, AccessorInfo.Key key, IfaceType ifaceType, String name) {
-                super(parent, key, ifaceType,name);
-
-            }
-
-            @Override
-            public void toText(String indent, Consumer<String> stringConsumer) {
-                stringConsumer.accept(indent + "mappable field " + key + ":" + type.iface);
-            }
-        }
 
         public static final class PrimitiveField extends AbstractPrimitiveField {
-
             PrimitiveField(IfaceType parent, AccessorInfo.Key key, Class<?> type, String name) {
                 super(parent, key, type, name);
 
@@ -450,32 +433,13 @@ public class Schema<T extends Buffer> {
             }
         }
 
-        public abstract static sealed class IfaceArray extends AbstractIfaceField permits  IfaceFieldControlledArray, IfaceFixedArray {
-            IfaceArray(IfaceType parent, AccessorInfo.Key key, IfaceType ifaceType, String name) {
-                super(parent, key, ifaceType, name);
-            }
-        }
 
-        public abstract static sealed class PrimitiveArray extends AbstractPrimitiveField permits  PrimitiveFieldControlledArray, PrimitiveFixedArray {
+        public abstract static sealed class PrimitiveArray extends AbstractPrimitiveField permits PrimitiveFieldControlledArray, PrimitiveFixedArray {
             PrimitiveArray(IfaceType parent, AccessorInfo.Key key, Class<?> type, String name) {
                 super(parent, key, type, name);
             }
         }
 
-        public static final class IfaceFixedArray extends IfaceArray {
-            public int len;
-
-            IfaceFixedArray(IfaceType parent, AccessorInfo.Key key, IfaceType ifaceType, String name, int len) {
-                super(parent, key, ifaceType, name);
-                this.len = len;
-            }
-
-            @Override
-            public void toText(String indent, Consumer<String> stringConsumer) {
-                stringConsumer.accept(indent + "array [" + len + "]");
-            }
-
-        }
 
         public static final class PrimitiveFixedArray extends PrimitiveArray {
             public int len;
@@ -488,24 +452,6 @@ public class Schema<T extends Buffer> {
             @Override
             public void toText(String indent, Consumer<String> stringConsumer) {
                 stringConsumer.accept(indent + "primitive array [" + len + "]");
-            }
-        }
-
-        public static final class IfaceFieldControlledArray extends IfaceArray {
-            List<ArrayLen> arrayLenFields;
-            int stride;
-            int contributingDims;
-
-            IfaceFieldControlledArray(IfaceType parent, AccessorInfo.Key key, IfaceType ifaceType, String name, List<ArrayLen> arrayLenFields, int stride) {
-                super(parent, key, ifaceType,name);
-                this.arrayLenFields = arrayLenFields;
-                this.stride = stride;
-                this.contributingDims = arrayLenFields.size();
-            }
-
-            @Override
-            public void toText(String indent, Consumer<String> stringConsumer) {
-                stringConsumer.accept(indent + name + "[" + key + ":" + type.iface + "] where len defined by " + arrayLenFields);
             }
         }
 
@@ -526,5 +472,70 @@ public class Schema<T extends Buffer> {
                 stringConsumer.accept(indent + name + "[" + key + ":" + type + "] where len defined by " + arrayLenFields);
             }
         }
+
+        public static abstract sealed class AbstractIfaceField extends FieldNode
+                permits FieldNode.IfaceArray, FieldNode.IfaceField {
+            public IfaceType ifaceType;
+
+            AbstractIfaceField(IfaceType parent, AccessorInfo.Key key, IfaceType ifaceType, String name) {
+                super(parent, key, name);
+                this.ifaceType = ifaceType;
+            }
+        }
+
+
+        public static final class IfaceField extends AbstractIfaceField {
+
+            IfaceField(IfaceType parent, AccessorInfo.Key key, IfaceType ifaceType, String name) {
+                super(parent, key, ifaceType, name);
+
+            }
+
+            @Override
+            public void toText(String indent, Consumer<String> stringConsumer) {
+                stringConsumer.accept(indent + "mappable field " + key + ":" + ifaceType.iface);
+            }
+        }
+
+
+        public abstract static sealed class IfaceArray extends AbstractIfaceField permits IfaceFieldControlledArray, IfaceFixedArray {
+            IfaceArray(IfaceType parent, AccessorInfo.Key key, IfaceType ifaceType, String name) {
+                super(parent, key, ifaceType, name);
+            }
+        }
+
+        public static final class IfaceFixedArray extends IfaceArray {
+            public int len;
+
+            IfaceFixedArray(IfaceType parent, AccessorInfo.Key key, IfaceType ifaceType, String name, int len) {
+                super(parent, key, ifaceType, name);
+                this.len = len;
+            }
+
+            @Override
+            public void toText(String indent, Consumer<String> stringConsumer) {
+                stringConsumer.accept(indent + "array [" + len + "]");
+            }
+        }
+
+        public static final class IfaceFieldControlledArray extends IfaceArray {
+            List<ArrayLen> arrayLenFields;
+            int stride;
+            int contributingDims;
+
+            IfaceFieldControlledArray(IfaceType parent, AccessorInfo.Key key, IfaceType ifaceType, String name, List<ArrayLen> arrayLenFields, int stride) {
+                super(parent, key, ifaceType, name);
+                this.arrayLenFields = arrayLenFields;
+                this.stride = stride;
+                this.contributingDims = arrayLenFields.size();
+            }
+
+            @Override
+            public void toText(String indent, Consumer<String> stringConsumer) {
+                stringConsumer.accept(indent + name + "[" + key + ":" + ifaceType.iface + "] where len defined by " + arrayLenFields);
+            }
+        }
+
+
     }
 }
