+++ hat/intellij/.idea/modules.xml
@@ -12,6 +12,7 @@
       <module fileurl="file://$PROJECT_DIR$/mandel.iml" filepath="$PROJECT_DIR$/mandel.iml" />
       <module fileurl="file://$PROJECT_DIR$/heal.iml" filepath="$PROJECT_DIR$/heal.iml" />
       <module fileurl="file://$PROJECT_DIR$/view.iml" filepath="$PROJECT_DIR$/view.iml" />
+      <module fileurl="file://$PROJECT_DIR$/life.iml" filepath="$PROJECT_DIR$/life.iml" />
       <module fileurl="file://$PROJECT_DIR$/squares.iml" filepath="$PROJECT_DIR$/squares.iml" />
       <module fileurl="file://$PROJECT_DIR$/experiments.iml" filepath="$PROJECT_DIR$/experiments.iml" />
       <module fileurl="file://$PROJECT_DIR$/violajones.iml" filepath="$PROJECT_DIR$/violajones.iml" />
