+++ hat/intellij/life.iml
@@ -1,0 +1,16 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<module type="JAVA_MODULE" version="4">
+  <component name="NewModuleRootManager" inherit-compiler-output="true">
+    <exclude-output />
+    <content url="file://$MODULE_DIR$/../hattricks/life">
+      <sourceFolder url="file://$MODULE_DIR$/../hattricks/life/src/main/java" isTestSource="false" />
+      <sourceFolder url="file://$MODULE_DIR$/../hattricks/life/src/main/resources" type="java-resource" />
+    </content>
+    <orderEntry type="inheritedJdk" />
+    <orderEntry type="sourceFolder" forTests="false" />
+    <orderEntry type="module" module-name="hat" />
+    <orderEntry type="module" module-name="backend_opencl" />
+    <orderEntry type="module" module-name="backend_cuda" />
+    <orderEntry type="module" module-name="backend_spirv" />
+  </component>
+</module>
